backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
Flask==3.1.2
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
MarkupSafe==3.0.3
oauthlib==3.3.1
openai==2.14.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
sniffio==1.3.1
SQLAlchemy==2.0.46
tqdm==4.67.1
tweepy==4.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
